Biblioteca_Comunal13/apps/usuarios/serializers.py
from ast import mod
from dataclasses import field
from pyexpat import model
from rest_framework import serializers
from .models import Usuario

class UsuarioSerializer(serializers.ModelSerializer):
    
    class Meta:
        model = Usuario
        fields = '__al__'