Biblioteca_Comunal13/requirements.txt
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
certifi==2025.4.26
cffi==1.17.1
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
Django==5.2.6
django-extensions==4.1
djangorestframework==3.16.1
executing==2.2.0
h11==0.16.0
idna==3.10
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
matplotlib-inline==0.1.7
MouseInfo==0.1.3
mypy_extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.2.6
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
prompt_toolkit==3.0.51
psutil==7.0.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.10.1
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
python-dateutil==2.9.0.post0
pytweening==1.2.0
pywin32==310
pyzmq==26.4.0
selenium==4.32.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.5.1
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
types-PyAutoGUI==0.9.3.20241230
types-PyScreeze==1.0.1.20250425
typing==3.7.4.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
wheel==0.45.1
wsproto==1.2.0
